==> Part4 - Feature extraction.ipynb <==
import matplotlib.pyplot as plt
import numpy as np
import cv2 
import pandas as pd
import shutil
import os 
import torch
import torchvision.models as models
import torchvision.transforms as transforms
from glob import glob
from tqdm import tqdm
from skimage.feature import hog, local_binary_pattern
from PIL import Image

resnet18 = models.resnet18(pretrained=True)
# Remove the fully connected layer to get feature vectors
resnet18 = torch.nn.Sequential(*list(resnet18.children())[:-1])
resnet18.eval()  # Set the model to evaluation mode

# Step 2: Define Image Preprocessing
preprocess = transforms.Compose([
    transforms.Resize((224, 224)),  # Resize image to 224x224
    transforms.ToTensor(),          # Convert image to tensor
    transforms.Normalize(mean=[0.485, 0.456, 0.406], std=[0.229, 0.224, 0.225])  # Normalize
])

dataDF = pd.read_csv("mixed_dataset_v1.csv")
print(dataDF.shape)
dataDF.head()

## Traditional methods

### Histogram of Oriented Gradients (HOG)

num_hog_features = 324
feature_columns = {f'hog_{i+1}': None for i in range(num_hog_features)}
hog_feature_df = pd.DataFrame(feature_columns, index=dataDF.index)

for idx, row in tqdm(dataDF.iterrows(), total=len(dataDF)):
    image_path = os.path.join("mixed_datasets", row["set"], "irex" if row["label"] == 1 else "others", row["filename"])

    # Load the image and convert to grayscale
    image = cv2.imread(image_path, cv2.IMREAD_GRAYSCALE)
    resized_image = cv2.resize(image, (64, 64))
    hog_features, hog_image = hog(
        resized_image,
        orientations=9,  # Number of gradient orientations
        pixels_per_cell=(16, 16),  # Size of cell
        cells_per_block=(2, 2),  # Number of cells per block
        block_norm='L2-Hys',  # Block normalization
        visualize=True,  # Output HOG image
        feature_vector=True,  # Return features as a vector
        )
    
    # Assign Features to DataFrame
    for i, feature in enumerate(hog_features):
        column_name = f"hog_{i+1}"
        if column_name in hog_feature_df.columns:  # Check if the column exists
            hog_feature_df.at[idx, column_name] = feature  # Assign feature to the appropriate column

hog_feature_df.head()

### Local Binary Patterns (LBP)

num_lbp_features = 25
feature_columns = {f'lbp_{i+1}': None for i in range(num_lbp_features)}
lbp_feature_df = pd.DataFrame(feature_columns, index=dataDF.index)

for idx, row in tqdm(dataDF.iterrows(), total=len(dataDF)):
    image_path = os.path.join("mixed_datasets", row["set"], "irex" if row["label"] == 1 else "others", row["filename"])

    # Load the image and convert to grayscale
    image = cv2.imread(image_path, cv2.IMREAD_GRAYSCALE)
    # Parameters for LBP
    radius = 3  # Radius of the circular neighborhood
    n_points = 8 * radius  # Number of points in the circular neighborhood
    lbp = local_binary_pattern(image, n_points, radius, method='uniform')
    n_bins = 25
    lbp_histogram, _ = np.histogram(lbp.ravel(), bins=n_bins, range=(0, n_bins))
    
    # Normalize the histogram
    lbp_histogram = lbp_histogram.astype("float")
    lbp_histogram /= (lbp_histogram.sum() + 1e-6)

    # Assign Features to DataFrame
    for i, feature in enumerate(lbp_histogram):
        column_name = f"lbp_{i+1}"
        if column_name in lbp_feature_df.columns:  # Check if the column exists
            lbp_feature_df.at[idx, column_name] = feature  # Assign feature to the appropriate column

lbp_feature_df.head()

## Extract Resnet18 features

num_re18_features = 512
feature_columns = {f'R18_{i+1}': None for i in range(num_re18_features)}
r18_feature_df = pd.DataFrame(feature_columns, index=dataDF.index)

for idx, row in tqdm(dataDF.iterrows(), total=len(dataDF)):
    image_path = os.path.join("mixed_datasets", row["set"], "irex" if row["label"] == 1 else "others", row["filename"])
    image = Image.open(image_path).convert("RGB")  # Open the image
    input_tensor = preprocess(image).unsqueeze(0)  # Preprocess and add batch dimension
    
    with torch.no_grad():
        features = resnet18(input_tensor).squeeze().numpy()  # Extract and convert to NumPy array

    # Assign Features to DataFrame
    for i, feature in enumerate(features):
        column_name = f"R18_{i+1}"
        if column_name in r18_feature_df.columns:  # Check if the column exists
            r18_feature_df.at[idx, column_name] = feature  # Assign feature to the appropriate column

r18_feature_df.head()

## Merge all features

finalDF = pd.concat([dataDF, hog_feature_df, lbp_feature_df, r18_feature_df], axis=1)
finalDF.head()

finalDF.to_csv("mixed_dataset_w_features_v1.csv", index=False)
