==> Part5 - Model Selection & Result Evaluation.ipynb <==
import matplotlib.pyplot as plt
import numpy as np
import cv2 
import pandas as pd
import shutil
import os 
import joblib

import pandas as pd
import numpy as np
import seaborn as sns
import matplotlib.pyplot as plt
from tqdm import tqdm
from sklearn.model_selection import train_test_split
from sklearn.impute import SimpleImputer
from sklearn.covariance import EllipticEnvelope
from matplotlib import rcParams
from sklearn.preprocessing import MinMaxScaler
from sklearn.tree import DecisionTreeRegressor
from sklearn.feature_selection import RFE
from sklearn.preprocessing import PolynomialFeatures
from sklearn.linear_model import LinearRegression
from sklearn.ensemble import RandomForestRegressor
from sklearn.model_selection import cross_val_score
from sklearn.svm import SVR
from sklearn.neural_network import MLPRegressor
from sklearn.metrics import r2_score, make_scorer
from tqdm import tqdm
import time
from sklearn.decomposition import PCA
from sklearn.preprocessing import StandardScaler
from sklearn.neural_network import MLPClassifier
from sklearn.metrics import accuracy_score
from sklearn.ensemble import RandomForestClassifier
from sklearn.model_selection import cross_val_score
from sklearn.svm import SVR
from sklearn.neural_network import MLPClassifier
from sklearn.metrics import r2_score, make_scorer
from tqdm import tqdm
from sklearn.neighbors import KNeighborsClassifier

dataDF = pd.read_csv("mixed_dataset_w_features_v1.csv")
print(dataDF.shape)
num_resnet_features = len([n for n in dataDF.columns if "R18" in n])
num_hog_features = len([n for n in dataDF.columns if "hog" in n])
num_lbp_features = len([n for n in dataDF.columns if "lbp" in n])
print(f"Num features: HOG ({num_hog_features}) | LBP ({num_lbp_features}) | Resnet18 ({num_resnet_features})")
dataDF.head()

## Data split

trainDF = dataDF[dataDF.set == "train"]
Ytrain = trainDF[["label"]]
Xtrain = trainDF.drop(columns=["filename", "set", "is_augmented", "label"])
validDF = dataDF[dataDF.set == "validation"]
Yval = validDF[["label"]]
Xval = validDF.drop(columns=["filename", "set", "is_augmented", "label"])
testDF = dataDF[dataDF.set == "test"]
Ytest = testDF[["label"]]
Xtest = testDF.drop(columns=["filename", "set", "is_augmented", "label"])
print(trainDF.shape, validDF.shape, testDF.shape)
Xtrain.shape, Ytrain.shape, Xval.shape, Yval.shape, Xtest.shape, Ytest.shape

## Prepocessing dataset

### Check missing values

missing_per_column = dataDF.isnull().sum()
print("Total missing value count:\n", sum(missing_per_column))

### Check range of features

dataDF.describe()

### Extract unaugmented train & valid data (no augmented samples in test set)

Xtrain_no_augmented = Xtrain[trainDF["is_augmented"] == 0]
Xval_no_augmented = Xval[validDF["is_augmented"] == 0]
Xtest_no_augmented = Xtest[testDF["is_augmented"] == 0]

Ytrain_no_augmented = Ytrain[trainDF["is_augmented"] == 0]
Yval_no_augmented = Yval[validDF["is_augmented"] == 0]
Ytest_no_augmented = Ytest[testDF["is_augmented"] == 0]

Xtrain_no_augmented.shape, Ytrain_no_augmented.shape, Xval_no_augmented.shape, Yval_no_augmented.shape, Xtest_no_augmented.shape, Ytest_no_augmented.shape

### Compress data with PCA

#### Normal data

scaler = StandardScaler()
scaler.fit(Xtrain)
Xtrain_scaled = scaler.transform(Xtrain)
Xval_scaled = scaler.transform(Xval)
Xtest_scaled = scaler.transform(Xtest)

pca = PCA(n_components=100)  # Specify the number of components to keep

# Fit and transform the data
Xtrain_scaled_pca100 = pca.fit_transform(Xtrain_scaled)
Xval_scaled_pca100 = pca.transform(Xval_scaled)
Xtest_scaled_pca100 = pca.transform(Xtest_scaled)

Xtrain_scaled_pca100.shape, Xval_scaled_pca100.shape, Xtest_scaled_pca100.shape

#### Unaugmented data

scaler_no_augmented = StandardScaler()
scaler_no_augmented.fit(Xtrain_no_augmented)
Xtrain_no_augmented_scaled = scaler_no_augmented.transform(Xtrain_no_augmented)
Xval_no_augmented_scaled = scaler_no_augmented.transform(Xval_no_augmented)
Xtest_no_augmented_scaled = scaler_no_augmented.transform(Xtest_no_augmented)

pca_no_augmented = PCA(n_components=100)  # Specify the number of components to keep

# Fit and transform the data
Xtrain_no_augmented_scaled_pca100 = pca_no_augmented.fit_transform(Xtrain_no_augmented_scaled)
Xval_no_augmented_scaled_pca100 = pca_no_augmented.transform(Xval_no_augmented_scaled)
Xtest_no_augmented_scaled_pca100 = pca_no_augmented.transform(Xtest_no_augmented_scaled)

Xtrain_no_augmented_scaled_pca100.shape, Xval_no_augmented_scaled_pca100.shape, Xtest_no_augmented_scaled_pca100.shape

## Choose the best preprocessing method

def validate_preprocessing_method(xtrain, xval, ytrain, yval):
    if type(xtrain) is pd.DataFrame:
        xtrain = xtrain.to_numpy()
    if type(xval) is pd.DataFrame:
        xval = xval.to_numpy()
        
    mlp_estimator = MLPClassifier(solver="sgd", random_state=42)
    mlp_estimator.fit(xtrain, ytrain['label'])
    ypreds = mlp_estimator.predict(xval)
    acc = accuracy_score(yval["label"], ypreds)
    start_time = time.time()
    for i in range(100):
        mlp_estimator.predict(np.reshape(xtrain[np.random.randint(0,100),:], (1,-1)))
    end_time = time.time()
    print(f"Inference time per sample: {(end_time-start_time)/100} seconds")
    print(f"Accuracy: {acc}")
    return acc

### Normal dataset

validate_preprocessing_method(Xtrain, Xval, Ytrain, Yval)

### Normal dataset - HOG features only

hog_columns = [f"hog_{i+1}" for i in range(num_hog_features)]
validate_preprocessing_method(Xtrain[hog_columns], Xval[hog_columns], Ytrain, Yval)

### Normal dataset - LBP features only

lbp_columns = [f"lbp_{i+1}" for i in range(num_lbp_features)]
validate_preprocessing_method(Xtrain[lbp_columns], Xval[lbp_columns], Ytrain, Yval)

### Normal dataset - Resnet18 features only

resnet_columns = [f"R18_{i+1}" for i in range(num_resnet_features)]
validate_preprocessing_method(Xtrain[resnet_columns], Xval[resnet_columns], Ytrain, Yval

### No augmented dataset

validate_preprocessing_method(Xtrain_no_augmented, Xval_no_augmented, Ytrain_no_augmented, Yval_no_augmented)

### Scaled dataset

validate_preprocessing_method(Xtrain_scaled, Xval_scaled, Ytrain, Yval)

### PCA dataset

validate_preprocessing_method(Xtrain_scaled_pca100, Xval_scaled_pca100, Ytrain, Yval)

### No augmented - Scaled dataset 

validate_preprocessing_method(Xtrain_no_augmented_scaled, Xval_no_augmented_scaled, Ytrain_no_augmented, Yval_no_augmented)

### No augmented - PCA dataset

validate_preprocessing_method(Xtrain_no_augmented_scaled_pca100, Xval_no_augmented_scaled_pca100, Ytrain_no_augmented, Yval_no_augmented)

## Choose suitable model family

Xtrain_scaled.shape, Xval_scaled.shape, Ytrain.shape, Yval.shape

def automatic_scoring(x, y, estimator):
    average_score = cross_val_score(estimator=estimator, X=x, y=y, cv=5, scoring='accuracy', error_score="raise").mean()
    return average_score

rf_estimator = RandomForestClassifier()
s = automatic_scoring(Xtrain_scaled, Ytrain["label"], rf_estimator)
print(f"Score of Random Forest Regression: {s}")

knn_estimator = KNeighborsClassifier()
s = automatic_scoring(Xtrain_scaled, Ytrain["label"], knn_estimator)
print(f"Score of KNN: {s}")

mlp_estimator = MLPClassifier()
s = automatic_scoring(Xtrain_scaled, Ytrain["label"], mlp_estimator)
print(f"Score of MLP: {s}")

## Find an optimal parameters for the best model family

## do grid search to find the best model parameters
# we will choose the param around the default values to check if the direction would be better (increasing or decreasing compared to the default ones) 
params_hidden_layer_sizes = [25, 35, 50, 75, 100, 200, 500]
params_learning_rates = ['constant', 'invscaling', 'adaptive']
params_learning_rate_inits = [0.0001, 0.001, 0.005, 0.0075, 0.01, 0.025, 0.05, 0.1]
params_max_iters = [50, 100, 200, 250, 300]
num_models = len(params_hidden_layer_sizes) * len(params_learning_rates) * len(params_learning_rate_inits) * len(params_max_iters)
print(f"Number of possible models: {num_models}")

## Train/val scheme
import warnings
warnings.filterwarnings("ignore")

best_params = None
best_acc = float('-inf')  # Initialize with negative infinity
start_time = time.time()
with tqdm(total=num_models) as pbar:
    for hls in params_hidden_layer_sizes:
        for lr in params_learning_rates:
            for lri in params_learning_rate_inits:
                for miter in params_max_iters:
                    model = MLPClassifier(solver="sgd", 
                                         hidden_layer_sizes=hls, 
                                         learning_rate=lr, 
                                         learning_rate_init=lri, 
                                         max_iter=miter,
                                         early_stopping=True,
                                         n_iter_no_change=4,
                                         random_state=42)
                    model.fit(Xtrain_scaled, Ytrain['label'])
                    ypreds = model.predict(Xval_scaled)
                    acc = accuracy_score(Yval["label"], ypreds)
                    
                    # Check if this set of hyperparameters is the best so far
                    if acc > best_acc:
                        best_acc = acc
                        best_params = {
                            'hidden_layer_sizes': hls,
                            'learning_rate': lr,
                            'learning_rate_init': lri,
                            'max_iter': miter
                        }
                    pbar.update(1)

end_time = time.time()

# Print the best hyperparameters and R-squared
print(f'Best Parameters: {best_params}')
print(f'Best accuracy: {best_acc:.4f}')
print(f"Total searching time: {round(end_time - start_time,4)} seconds")

## Validate on test set

model = MLPClassifier(solver="sgd", 
                     hidden_layer_sizes=best_params['hidden_layer_sizes'], 
                     learning_rate=best_params['learning_rate'], 
                     learning_rate_init=best_params['learning_rate_init'], 
                     max_iter=best_params['max_iter'],
                     early_stopping=True,
                     n_iter_no_change=4,
                     random_state=42)
model.fit(Xtrain_scaled, Ytrain['label'])
ytrainpreds = model.predict(Xtrain_scaled)
yvalpreds = model.predict(Xval_scaled)
ypreds = model.predict(Xtest_scaled)
acctrain = accuracy_score(Ytrain["label"], ytrainpreds)
accval = accuracy_score(Yval["label"], yvalpreds)
acc = accuracy_score(Ytest["label"], ypreds)
print(f'Train/val/test scoring comparison in case using only the training part:')
print(f'Accuracy training score: {acctrain:.2f}')
print(f'Accuracy validating score: {accval:.2f}')
print(f'Accuracy testing score: {acc:.2f}')

## Export the best model and some related material

joblib.dump(model, 'mlp_classifier_PblHogResnet18_861Features_model_v02.pkl')

joblib.dump(scaler, 'scaler.pkl')
